+++ Assets/scripts/Trees/SunFlower.cs
@@ -52,7 +52,7 @@ public class SunFlower : ActionOfTreeGetSun
         // Đảm bảo rằng đối tượng đến đúng vị trí cuối cùng
         targetTransform.position = targetPosition;
     }
-    public override void Hit(int damage)
+    public override void Hit(float damage)
     {
         Debug.Log("Bi can Bi Can");
         hp -= damage;
